backend/src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Evgeniy Goncharuck <iterlace@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
pydantic = "^1.8.2"
tenacity = "^8.0.1"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
mypy = "^0.910"
black = "^21.7b0"
isort = "^5.9.2"
autoflake = "^1.4"
flake8 = "^3.9.2"
pytest-cov = "^2.12.1"
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
